chore: change community examples to advanced and refactor custom_url.py as downsample.py
