+++ Examples/README.md
@@ -18,6 +18,7 @@ Richer examples can be found in the `Examples/write` directory.
 
    * `batching.py` - shows how to make use of the _batching_ API for writing long-running data sets.   
    * `fileimport.py` - shows how to import data to an Influx database directly from a number of other standard database formats.
+       * To refresh the source data use `Examples/write/source_data/updater.py` 
    * `handle_http_error.py` - shows error handling on writes.
    * `pandas_write.py` - shows how to write pandas dataframes directly to an Influx database.
    * `writeoptions.py` - shows the core options API for writes. (TODO more specific?  See point 7.1 below)
@@ -29,6 +30,16 @@ Basic examples can be found in the `Examples/core` directory.
    * `basic_query.py` - shows the essentials of querying and Influxdb database.
    * `basic_ssl.py` - shows how to handle SSL and TLS certificates. (TODO see point 7.2 below.)
    * `timeouts.py` - shows how to set and leverage timeout values.
+
+Richer examples can be found in the `Examples/query` directory.
+
+   * `flight_options.py` - shows how to set options on the query transport.
+   * `handle_query_error.py` - shows basic error handling.
+   * `query_async.py` - shows basic usage of the asynchronous query API. 
+   * `query_modes.py` - when making a query, different modes return data in different types of structures.  This example shows which modes return which structured formats.
+   * `query_with_middleware.py` - when making a query it is possible to insert special headers into the HTTP layer using middleware.  This example shows how to do this.
+
+### Advanced examples
 
 ## Refactoring Notes 
 TODO - delete this section as examples take shape and before creating PR. 
@@ -49,6 +60,24 @@ TODO - delete this section as examples take shape and before creating PR.
 2. Keep `file-import` as single file, with source data updatable - see `updater.py` __DONE__
 3. Keep one simple example of jupiter notebook. __DONE__
 4. Decide what to do with `./community` examples.
+   * `custom_url.py` 
+       * Doesn't seem to do what is on the tin.
+       * Seems more like a _down sampling_ example
+       * Doesn't query Influxdb directly but down samples from remote CSV
+       * TODO 
+           * rename to _down sampling_ or similar
+           * use initial query
+           * move to `./advanced`
+           * remove dependency on `githubusercontent`
+   * `database_transfer.py` 
+       * Simply copies data from one bucket to another
+       * Useful base example
+       * Current illustrative only - doesn't look functional without some undocumented setup
+       * To work requires `dbfrom` and measurement `airSensors` to exist.
+           * Note - are camel caps not sometimes problematic in Influx? (seem to recall encountering problem with them in the past)
+       * TODO
+           * move to `./advanced`
+           * make functional
 5. Root examples
    1. `basic_ssl_examle.py` to `./core` __DONE__
    2. `batching_example.py` to `./write` __DONE__
@@ -70,9 +99,12 @@ TODO - delete this section as examples take shape and before creating PR.
    1. Some examples show leveraging internal **kwargs like `data_frame_measurement_name` or `data_frame_tag_columns`
        - Makes sense to expose these in _advanced_ examples. (e.g. pandas examples... )
        - Perhaps though should encourage the use of a simpler standard API that hides them
+   2. Remove dependencies on remote `../githubusercontent/../*.csv` 
    2. Leverage `config.py` in all examples
    3. Prefer using Influxdb3 Core by default.  But also document possibility of using other products.
 7. Enhancements
    1. `writeoptions.py` - does not show much in the way of setting options.
    2. `basic_ssl.py` - review. Seems to only show handling SSL handshake failures.
 8. Verify all refactored examples are working
+   *. __NOTE__ - If making an _illustrative_ example functional _out-of-the-box_ leads to too much distractive information being added, leave the example as _illustrative only_ and add a comment that it is for purposes of illustration.  However, make sure the illustrative example is still working in a concrete implementation. (e.g. `query_with_middleware.py`)
+   *. TODO In comments, mark examples as either _illustrative_ or _functional_
