+++ Examples/query/query_with_middleware.py
@@ -1,6 +1,6 @@
 from pyarrow import flight
 
-from config import Config
+from Examples.config import Config
 from influxdb_client_3 import InfluxDBClient3, flight_client_options
